refactor: 의존성 확인을 위한 패키지 이동
