+++ backend/src/main/java/com/yigongil/backend/domain/member/MemberService.java
@@ -1,9 +1,7 @@
-package com.yigongil.backend.application;
+package com.yigongil.backend.domain.member;
 
-import com.yigongil.backend.domain.member.Member;
-import com.yigongil.backend.domain.member.MemberRepository;
-import com.yigongil.backend.domain.member.Nickname;
 import com.yigongil.backend.domain.study.Study;
+import com.yigongil.backend.domain.study.StudyService;
 import com.yigongil.backend.domain.studymember.StudyMember;
 import com.yigongil.backend.domain.studymember.StudyMemberRepository;
 import com.yigongil.backend.exception.MemberNotFoundException;
