+++ backend/src/main/java/com/yigongil/backend/domain/round/MustDoService.java
@@ -1,8 +1,6 @@
-package com.yigongil.backend.application;
+package com.yigongil.backend.domain.round;
 
 import com.yigongil.backend.domain.member.Member;
-import com.yigongil.backend.domain.round.Round;
-import com.yigongil.backend.domain.round.RoundRepository;
 import com.yigongil.backend.exception.RoundNotFoundException;
 import com.yigongil.backend.request.MustDoUpdateRequest;
 import org.springframework.stereotype.Service;
