+++ backend/src/test/java/com/yigongil/backend/application/StudyServiceTest.java
@@ -4,9 +4,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
+import com.yigongil.backend.domain.certification.CertificationService;
+import com.yigongil.backend.domain.feedpost.FeedService;
 import com.yigongil.backend.domain.member.Member;
 import com.yigongil.backend.domain.round.RoundRepository;
 import com.yigongil.backend.domain.study.StudyRepository;
+import com.yigongil.backend.domain.study.StudyService;
 import com.yigongil.backend.domain.studymember.StudyMemberRepository;
 import com.yigongil.backend.exception.StudyNotFoundException;
 import com.yigongil.backend.fixture.MemberFixture;
