+++ backend/src/test/java/com/yigongil/backend/fake/FakeController.java
@@ -1,9 +1,9 @@
 package com.yigongil.backend.fake;
 
-import com.yigongil.backend.application.StudyService;
 import com.yigongil.backend.config.auth.JwtTokenProvider;
 import com.yigongil.backend.domain.member.Member;
 import com.yigongil.backend.domain.member.MemberRepository;
+import com.yigongil.backend.domain.study.StudyService;
 import com.yigongil.backend.response.TokenResponse;
 import java.time.LocalDate;
 import java.util.Optional;
