+++ backend/src/test/java/com/yigongil/backend/ui/StudyControllerTest.java
@@ -9,13 +9,14 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yigongil.backend.application.CertificationService;
-import com.yigongil.backend.application.FeedService;
-import com.yigongil.backend.application.MustDoService;
-import com.yigongil.backend.application.StudyService;
 import com.yigongil.backend.config.auth.AuthContext;
 import com.yigongil.backend.config.auth.JwtTokenProvider;
+import com.yigongil.backend.domain.certification.CertificationService;
+import com.yigongil.backend.domain.feedpost.FeedService;
 import com.yigongil.backend.domain.member.MemberRepository;
+import com.yigongil.backend.domain.round.MustDoService;
+import com.yigongil.backend.domain.study.StudyController;
+import com.yigongil.backend.domain.study.StudyService;
 import com.yigongil.backend.fixture.MemberFixture;
 import com.yigongil.backend.request.StudyUpdateRequest;
 import com.yigongil.backend.ui.exceptionhandler.InternalServerErrorMessageConverter;
